3.4
Полностью перевел с redux на mobx. Навел порядок в компонентах. Разбил сложные компоненты на простые

+++ src/Components/Mobx/Thunks.ts
@@ -1,3 +1,5 @@
+import { toJS } from 'mobx';
+import { observer } from 'mobx-react-lite';
 import { CardsAPI } from "../../Api/Api"
 import { U_T } from "../Redux/ReduxTypes"
 import LocalStorage from "./LocalStorage"
@@ -37,25 +39,42 @@ export const PushData_Thunk = async (SetMessageError: any) => {
         LocalStorage.toggleLoading(false)
     } catch(e: any) {
         if(e.response.status === 403) {
-            SetMessageError('Почта неактивирована. Доступны не все функции')
+            SetMessageError('Почта не подтверждена. Доступны не все функции')
         }
     } finally {
         LocalStorage.toggleLoading(false)
     }
 }
-export const ChangeCard_Thunk = async (cardID: number, text: string, groupsIDs: Array<number>, isCompleted: boolean) => {
-    LocalStorage.changeCard(cardID, text, groupsIDs, isCompleted)
-    LocalStorage.updateCurrentCards()
+export const ChangeCard_Thunk = (cardID: number, text: string, groupsIDs: Array<number>, isCompleted: boolean, SetMessageError: any) => {
+    if(LocalStorage.isActivated === 0) {
+        SetMessageError('Почта не подтверждена. Доступны не все функции')
+    } else {
+        LocalStorage.changeCard(cardID, text, groupsIDs, isCompleted)
+        LocalStorage.updateCurrentCards()
+    }
 }
-export const SwitchCompleteCard_Thunk = async (cardID: number) => {
+export const SwitchCompleteCard_Thunk = (cardID: number, SetMessageError: any) => {
+    if(!LocalStorage.isActivated) {
+        return SetMessageError('Почта не подтверждена. Доступны не все функции')
+    }
     LocalStorage.switchCompleteCard(cardID)
     LocalStorage.updateCurrentCards()
 }
-export const DeleteCard_Thunk = async (cardID: number) => {
+export const DeleteCard_Thunk = async (cardID: number, SetMessageError: any) => {
+    if(!LocalStorage.isActivated) {
+        return SetMessageError('Почта не подтверждена. Доступны не все функции')
+    }
     LocalStorage.deleteCard(cardID)
     LocalStorage.updateCurrentCards()
 }
-export const AddCard_Thunk = async (cardID: number, text: string, groupsIDs: Array<number>, isCompleted: boolean) => {
+export const AddCard_Thunk = (cardID: number, text: string, groupsIDs: Array<number>, isCompleted: boolean, SetMessageError: any) => {
+    if(!LocalStorage.isActivated) {
+        return SetMessageError('Почта не подтверждена. Доступны не все функции')
+    }
     LocalStorage.addNewCard(cardID, text, groupsIDs, isCompleted)
+    LocalStorage.updateCurrentCards()
+}
+export const SwitchCardGroup_Thunk = (groupID: number) => {
+    LocalStorage.changeCurrentCardGroupID(groupID)
     LocalStorage.updateCurrentCards()
 }
